+++ data/effects_data/.gitignore
