+++ data/effects_data/proguard-rules.pro
