+++ data/effects_data/src/androidTest/java/com/example/data/ExampleInstrumentedTest.kt
