+++ data/effects_data/src/main/AndroidManifest.xml
