+++ data/effects_data/src/main/java/com/cerebus/tokens/data/effects_data/storage/EffectsStorage.kt
@@ -1,4 +1,4 @@
-package com.cerebus.tokens.data.storage
+package com.cerebus.tokens.data.effects_data.storage
 
 /**
  * [EffectsStorage] - an interface for storage of effects data:
