+++ data/effects_data/src/main/java/com/cerebus/tokens/data/effects_data/storage/EffectsStorageImpl.kt
@@ -1,4 +1,4 @@
-package com.cerebus.tokens.data.storage
+package com.cerebus.tokens.data.effects_data.storage
 
 import android.content.Context
