+++ data/tokens_data/src/main/java/com/cerebus/tokens/data/tokens_data/storage/TokensStorage.kt
@@ -1,4 +1,4 @@
-package com.cerebus.tokens.data.storage
+package com.cerebus.tokens.data.tokens_data.storage
 
 /**
  * [TokensStorage] - an interface for storage of tokens data:
