+++ data/tokens_data/src/main/java/com/cerebus/tokens/data/tokens_data/storage/TokensStorageImpl.kt
@@ -1,4 +1,4 @@
-package com.cerebus.tokens.data.storage
+package com.cerebus.tokens.data.tokens_data.storage
 
 import android.content.Context
